Verlassen ohne Speichern abfrage (#784)
* Verlassen ohne Speichern abfrage

* AbstractJVereinControl zu Interface Savable geändert, fill zu
repareStore umbenannt

* AbstractJVereinVIew nach AbstractDetailView umbenannt

* Typo

* History erhalten bei abrechen des verlassens

* Speichern Dialog it Speichern Verlassen und Abbrechen

* Typo

* DBUpdate, Auch Eingenschaften, Zusatzfelder und sekundäre Beitragsgrupe
überwachen

* Fehler behoben

* null check

* Importe "" statt null setzen.

* Update weitere Spalten

* Update umbenannt

* Update angepasst

* Beitragsgruppe auch Betrag auf änderungen prüfen

* Änderungstest bei Splibuchungen entfernt

* Änderungsüberwachung auch bei Steuern

+++ src/de/jost_net/JVerein/gui/action/FormularDeleteAction.java
@@ -86,7 +86,7 @@ public class FormularDeleteAction implements Action
         return;
       }
       // Do not delete a form if it is linked to another
-      Integer formlink = f.getFormlink();
+      Long formlink = f.getFormlink();
       if (formlink > 0)
       {
         Formular fo = (Formular) Einstellungen.getDBService().createObject(
